move deadlineNanos
